Extract getImports into its own file
